changetheview_bot.js
For future expansions
