+++ bot.js
@@ -2,7 +2,7 @@
 const low = require('lowdb')
 const FileSync = require('lowdb/adapters/FileSync')
 
-const adapter = new FileSync('db.json')
+const adapter = new FileSync('./db.json')
 const db = low(adapter)
 db.defaults({users: [], usercount: 0, servers: [], plot_no: 0}).write();
 
@@ -13,9 +13,11 @@ const bot = new Discord.Client();
 // Prerequisites
 var logger = require('winston');
 var auth = require('./auth.json');
-var plotly = require('plotly')('zeklewa','sC8bIL6vurb1J24M7zNN');
+var global.plotly = require('plotly')('zeklewa','sC8bIL6vurb1J24M7zNN');
 var fs = require('fs');
 var request = require('request');
+var ytdl = require('ytdl-core');
+var search = require('youtube-search');
 
 // External functions
 var rqfunc = require('./required');
@@ -24,6 +26,10 @@ var rqfunc = require('./required');
 var start_up_time;
 var default_channel = "";
 var cur_prefix = "$";
+var servers = {}; // Key = server, Value = music queue for that server
+
+//External function
+var controller = require('./custom_modules/controller.js');
 
 // Configure logger settings
 logger.remove(logger.transports.Console);
@@ -31,6 +37,27 @@ logger.add(logger.transports.Console, {
     colorize: true
 });
 logger.level = 'debug';
+
+// Music function
+function play(connection, message, toChannel)
+{
+    var server = servers[message.guild.id];
+
+    toChannel.send("*Now playing:* `" + server.queue[0].title + "`.");
+    server.dispatcher = connection.playStream(ytdl(server.queue[0].link, {filter: 'audioonly'}));
+
+    server.queue.shift();
+
+    server.dispatcher.on("end", function(){
+        if (server.queue[0]) play(connection, message, toChannel);
+        else 
+        {
+            connection.disconnect();
+            toChannel.send('*Queue concluded.*');
+        } 
+    });
+    
+}
 
 // Login with specified token
 bot.login(auth.token);
@@ -220,12 +247,12 @@ bot.on('message', message => {
     if (args.length > 1)
     {
         para = args[1];
-    }        
+    }      
 
     args = args.splice(1);
 
     // If default_channel has not been set, reply to current channel
-    var toChannel = message.channel;
+    var global.toChannel = message.channel;
 
     if (default_channel){
         toChannel = default_channel;
@@ -244,121 +271,26 @@ bot.on('message', message => {
         mess += "The current prefix is* `" + cur_prefix + "`.";
         toChannel.send(log_mess + mess);
     }
+
+    //get the default arguments
+    var String_Args = "";
+    for (int i=0; i<args.length; i++){
+        String_Args = String_Args + args[i];
+        if (i<args.length) String_Args = String_Args + " ";
+    }
+
     // Default prefix settings
     else if (msg_content.substring(0, cur_prefix.length) == cur_prefix)
     {
         switch(cmd) {
             // Change prefix
             case 'prefix':
-                if (args.length == 0)
-                {
-                    toChannel.send(log_mess + "Usage: $prefix [desired_prefix]. You can use $prefix show to show the default prefix.*");
-                }
-                else
-                {
-                    if (message.member.permissions.has('MANAGE_GUILD') || (author_id == "176963972044423168"))
-                    {
-                        var req_prefix = msg_content.substring(cur_prefix.length);
-                        req_prefix = req_prefix.substr(req_prefix.indexOf(" ") + 1);
-                        db.get('servers').find({'guild_id' : guild_id}).assign({'prefix' : req_prefix}).write();
-                        toChannel.send(log_mess + "The prefix has been set to* `" + req_prefix + "`.");
-                        cur_prefix = req_prefix;
-                    }
-                    else
-                    {
-                        toChannel.send(log_mess + "You are not authorized to change the prefix!*");
-                    }
-                }
+                controller.changePrefix(String_Args);
             break;
 
             // uncen stuff
             case 'uncen':
-                if (args.length == 0)
-                {
-                    toChannel.send(log_mess + "Usage: $uncen add/remove [channel1] [channel2] [channel3] ...* to add/remove channels.");
-                }
-                else
-                {
-                    if (args[0] == 'show')
-                    {
-                        toChannel.send(log_mess + "The following channels are uncensored:* `" + uncen_channels + "`");
-                    }
-                    else if (args[0] == 'add')
-                    {
-                        if (message.member.permissions.has('MANAGE_CHANNELS') || (author_id == "176963972044423168"))
-                        {
-                            if (args.length == 1)
-                            {
-                                toChannel.send(log_mess + "You have to specify your channels!*");
-                            }
-                            else
-                            {
-                                queue_channels = args.slice(1);
-                                // toChannel.send(log_mess + "You are attempting to uncensor these channels:* `" + queue_channels +"`");
-
-                                var iter;
-                                for (iter = 0; iter < queue_channels.length; iter++)
-                                {
-                                    uncen_channel = queue_channels[iter];
-
-                                    if (!uncen_channels.includes(uncen_channel))
-                                    {
-                                        uncen_channels.push(uncen_channel);
-                                    }
-                                }
-
-                                db.get('servers').find({'guild_id' : guild_id}).assign({'uncen' : uncen_channels}).write();
-                                toChannel.send(log_mess + "The list of uncensored channels have been updated.*");
-                            }
-                        }
-                        else
-                        {
-                            toChannel.send(log_mess + "You are not authorized to edit channels!*");
-                        }           
-                    }
-                    else if (args[0] == 'remove')
-                    {
-                        if (message.member.permissions.has('MANAGE_CHANNELS') || (author_id == "176963972044423168"))
-                        {
-                            if (args.length == 1)
-                            {
-                                toChannel.send(log_mess + "You have to specify your channels!*");
-                            }
-                            else
-                            {
-                                queue_channels = args.slice(1);
-                                // toChannel.send(log_mess + "You are attempting to re-censor these channels:* `" + queue_channels + "`");
-
-                                if (queue_channels[0] == "*")
-                                {
-                                    // Remove all
-                                    db.get('servers').find({'guild_id' : guild_id}).assign({'uncen' : []}).write();
-                                    toChannel.send(log_mess + "The list of uncensored channels have been cleared.*");
-                                }
-                                else
-                                {
-                                    var iter;
-                                    for (iter = 0; iter < queue_channels.length; iter++)
-                                    {
-                                        uncen_channel = queue_channels[iter];
-
-                                        if (uncen_channels.includes(uncen_channel))
-                                        {
-                                            uncen_channels.splice(uncen_channels.indexOf(uncen_channel), 1);
-                                        }
-                                    }
-
-                                    db.get('servers').find({'guild_id' : guild_id}).assign({'uncen' : uncen_channels}).write();
-                                    toChannel.send(log_mess + "The list of uncensored channels have been updated.*");
-                                }        
-                            }
-                        }   
-                        else
-                        {
-                            toChannel.send(log_mess + "You are not authorized to edit channels!*");
-                        }                  
-                    }
-                }
+                controller.uncen(String_Args);
             break;
 
             // !ping
@@ -368,237 +300,144 @@ bot.on('message', message => {
 
             // Display uptime
             case 'uptime':
-                var uptime = + new Date() - start_up_time;
-                console.log(uptime);
-                toChannel.send(log_mess + "I have been up for " + rqfunc.timeConverter(uptime) + ". Beep boop.*");
+                controller.uptime();
             break;
 
             // Set default channel
             case 'defchan':
-                if (para == "")
-                {
-                    toChannel.send(log_mess + "Usage: $defchan channel_name_here. Use $defchan default to clear the default channel and $defchan show to show the default channel.*");
-                }
-
-                else if (para == "default")
-                {
-                    if ((message.member.permissions.has('MANAGE_CHANNELS')) || (author_id == "176963972044423168"))
-                    {
-                        toChannel.send(log_mess + "Default channel has been cleared.*");
-                        db.get('servers').find({'guild_id': guild_id}).assign({ "defchan": "" }).write();
-                    }
-                    else
-                        toChannel.send(log_mess + "You are not authorized to edit channels!.*")
-                }
-
-                else if (para == "show")
-                {
-                    if (!default_channel)
-                        toChannel.send(log_mess + "No default channel is currently set on this server.*");
-                    else
-                        toChannel.send(log_mess + "The default channel is currently <#" + default_channel.id + ">*");
-                }
-
-                else
-                {
-                    if (message.member.permissions.has('MANAGE_CHANNELS') || (author_id == "176963972044423168"))
-                    {
-                        var channel_found = false;
-                        var mess;
-
-                        found_channel = message.guild.channels.find("name", para);
-
-                        if (found_channel)
-                        {
-                            mess = 'Channel found! Default channel has been set.*';
-                            default_channel = found_channel;
-                            db.get('servers').find({'guild_id': guild_id}).assign({ "defchan": default_channel.id }).write();
-                        }
-
-                        else mess = 'Channel not found. Try again.*';
-                        toChannel.send(log_mess + mess);
-                    }
-                    else
-                        toChannel.send(log_mess + "You are not authorized to edit channels!.*")
-                }
-            break;
-            // Generate random imgur image
-            case 'rdimgur':
-                var iter;
-                var not_found = true;
-
-                for (iter = 0; iter < 20; iter++)
-                {
-                    var imgur_link = rqfunc.genImgur();
-                    if (rqfunc.imageExists(imgur_link)){
-                        toChannel.send(log_mess + imgur_link + "*");
-                        not_found = false;
-                        break;
-                    }
-                }
-
-                if (not_found)
-                {
-                    toChannel.send(log_mess + "Fuck imgur.");
-                }
+                controller.defchan(String_Args);
             break;
 
             // Banned words
             case 'bword':
-                var cur_bwords = db.get('servers').find({'guild_id' : guild_id}).get('banned_words').value();
-                if ((args.length >= 2) && ((args[0] == 'add') || (args[0] == 'remove')))
+                controller.bword(String_Args);
+            break;
+
+            case 'play':
+                // Check if queue exists
+                if (args.length == 0)
                 {
-                    var key_word = msg_content.substring(1);
-                    key_word = key_word.substr(key_word.indexOf(" ") + 1);
+                    toChannel.send(log_mess + 'Usage: $play [youtube search query]*, $skip to skip the current song, $stop to stop playing and $queue to display the current queue.*');
+                    return;
+                }
 
-                    // Used to remove "add"/"remove" in string
-                    key_word = key_word.substr(key_word.indexOf(" ") + 1);
+                if (!message.member.voiceChannel)
+                {
+                    toChannel.send(log_mess + 'You must be connected to a voice channel!*');
+                    return;
+                }
 
-                    //var key_word = args[1];
-                    console.log(key_word);
+                if (!servers[message.guild.id])
+                {
+                    servers[message.guild.id] = {queue : []};
+                }
 
-                    if (args[0] == 'show')
-                        toChannel.send(log_mess + "`Currently banned words: " + cur_bwords + "`");
-                    // If word exists
+                // Get query
+                var query = msg_content.substring(1);
+                query = query.substr(query.indexOf(" ") + 1);
 
-                    if (cur_bwords.includes(key_word))
+                var opts = {
+                    maxResults: 5,
+                    key: auth.youtubeAPI
+                };
+ 
+                // Process query here first -> push query to queue
+                search(query, opts, function(err, results) {
+                    // Select first object from search
+                    if (results)
                     {
-                        if (message.member.permissions.has('MANAGE_MESSAGES') || (author_id == "176963972044423168"))
+                        var result_song = results[0];
+
+                        // Push to queue
+                        toChannel.send(log_mess + 'The song* `' + result_song.title + '` *has been added to the queue*');
+                        var server = servers[message.guild.id];
+                        server.queue.push(result_song);          
+                            
+                        if (!message.guild.voiceConnection)
                         {
-                            if (args[0] == 'add')
-                            {
-                                toChannel.send(log_mess + "The word '" + key_word + "' already exists!*");
-                            }
-                            else
-                            {
-                                cur_bwords.splice(cur_bwords.indexOf(key_word), 1);
-                                db.get('servers').find({'guild_id' : guild_id}).assign({'banned_words' : cur_bwords}).write();
-                                toChannel.send(log_mess + "The word '" + key_word + "' has been deleted from the ban list.*");
-                            }
+                            message.member.voiceChannel.join().then(connection => {
+                                play(connection, message, toChannel);
+                            })
                         }
-                        else
-                            toChannel.send(log_mess + "You are not authorized to edit messages!*");
-                    }   
-                    // If word doesn't exist
+                    }
                     else
                     {
-                        if (message.member.permissions.has('MANAGE_MESSAGES') || (author_id == "176963972044423168"))
-                        {
-                            if (args[0] == 'add')
-                            {
-                                if (key_word == "*")
-                                {
-                                    toChannel.send(log_mess + "Uh oh, you're not allowed to do that you cheeky bastard.*"); 
-                                }
-                                else
-                                {
-                                    cur_bwords.push(key_word);
-                                    db.get('servers').find({'guild_id' : guild_id}).assign({'banned_words' : cur_bwords}).write();
-                                    toChannel.send(log_mess + "The word '" + key_word + "' has been banned.*"); 
-                                } 
-                            }
-                            else
-                            {
-                                if (key_word == "*")
-                                {
-                                    toChannel.send(log_mess + "The ban list has been cleared.*"); 
-                                    db.get('servers').find({'guild_id' : guild_id}).assign({'banned_words' : []}).write();
-                                }
-                                else
-                                    toChannel.send(log_mess + "The word '" + key_word + "' doesn't exist in the database!*");
-                            }
-                        }
-                        else
-                            toChannel.send(log_mess + "You are not authorized to edit messages!*");
+                        toChannel.send(log_mess + "Sorry, I couldn't find anything with that query.*");
                     }
+                if(err) return console.log(err); });
+            break;
+
+            // Skip current song
+            case 'skip':
+                var server = servers[message.guild.id];
+                if (server.dispatcher) server.dispatcher.end();
+            break;
+
+            // Stop playing
+            case 'stop':
+                var server = servers[message.guild.id];
+                if (message.guild.voiceConnection) 
+                {
+                    bot.voiceConnections.get(guild_id).channel.leave();
                 }
                 else
-                {
-                    if ((args.length == 1) && (args[0] == 'show'))
-                    {
-                        toChannel.send(log_mess + "`Currently banned words: " + cur_bwords + "`");
-                    }
-                    // Display all banned words and documentation for bword
-                    else toChannel.send(log_mess + "Usage: $bword add [banned_word], $bword remove [banned_word], $bword show.*");
-                }
+                    toChannel.send(log_mess + 'I am not connected to a channel!*');
             break;
-            // Plot
-            case 'plot':
-                try {
-                    // args = msg_content.substring(1).split(' ');
-                    console.log(args);
-                    if ((args.length < 3) || (args.length > 6))
-                        toChannel.send(log_mess + "Usage: $plot [function#], [begin#], [end#], [no. of points]\n For special functions in the JS math library, use prefix $. Ex: $sin, $cos, etc. \n Still under development, will have more functionalities in the future.*");
+
+            case 'volume':
+                var server = servers[message.guild.id];
+                if (args.length != 1)
+                {
+                    toChannel.send(log_mess + 'Usage: $volume [desired volume] [0-200](%) (100% being the normal perceived volume)). Use $volume show to display the current volume level.*');
+                    return;
+                } 
+                else
+                {
+                    if (args[0] == 'show')
+                        toChannel.send(log_mess + "The current volume is set at `" + server.dispatcher.volumeLogarithmic*100 + "%`.*");
+                    else if (isNaN(args[0]))
+                        toChannel.send(log_mess + "Please provide a number for the desired volume!*");
                     else
                     {
-                        db.update('plot_no', n => n + 1).write();
-                        var plot_no = db.get('plot_no').value();
-
-                        var func = args[0].replace('$', 'Math.').replace('^', '**');
-                        var begin = Number(args[1]);
-                        var end = Number(args[2]);
-                        var steps;
-
-                        if (begin > end)
+                        var set_volume = Number(args[0])/100;
+                        if ((set_volume >= 2) || (set_volume <= 0))
                         {
-                            var temp = begin;
-                            begin = end;
-                            end = temp;
+                            toChannel.send(log_mess + "Please provide a number within the specified bounds!*");
                         }
-
-                        im_w = 600;
-                        im_h = 600;
-                        steps = 100;
-
-                        if (args.length == 4)
-                        {
-                            steps = Number(args[3]);
-                        }
-
-                        var f = new Function("x", "return " + func);
-
-                        console.log(begin + " " + end + " " + steps);
-
-                        var xs = [];
-                        var ys = [];
-
-                        var k;
-                        for (k = 0; k < steps; k++)
-                        {
-                            x = begin + (end - begin)*k*1.0/steps;
-                            y = f(x);
-                            xs.push(x);
-                            ys.push(y);
-                        }
-
-                        var trace1 = {
-                          x: xs,
-                          y: ys,
-                          type: "scatter"
-                        };
-                         
-                        var figure = { 'data': [trace1] };
-                         
-                        var imgOpts = {
-                            format: 'png',
-                            width: im_w,
-                            height: im_h
-                        };
-                         
-                        var plot_path = 'Plots/' + plot_no +'.png';
-
-                        plotly.getImage(figure, imgOpts, function (error, imageStream) {
-                            if (error) return console.log (error);
-                            var fileStream = fs.createWriteStream(plot_path);
-                            fileStream.on('close', function(){
-                                toChannel.send(log_mess + "Plot created.*", {files: ['Plots/' + plot_no +'.png']});
-                            })
-                            imageStream.pipe(fileStream);
-                        });
+                        server.dispatcher.setVolumeLogarithmic(set_volume);
+                        toChannel.send(log_mess + 'The volume has been set to* `' + Number(args[0]) + '%`.');
+                        //toChannel.send(log_mess + 'Usage: $volume [desired volume] [0-200](%) (100% being the normal perceived volume). Use $volume show to display the current volume level.*');
                     }
-                } catch(err) {
-                    toChannel.send(log_mess + "Wrong syntax error!* `" + err + '`');
+                }           
+            break;
+
+            // Add song to queue
+            case 'queue':
+                var server = servers[message.guild.id];
+                if (args.length == 0)
+                {
+                    if (server.queue.length)
+                    {
+                        var string_mess = "*Current songs in the queue:* \n" + "```";
+                        var i;
+                        for (i = 0; i < server.queue.length; i++)
+                        {
+                            string_mess += (i + 1) + ". " + server.queue[i].title + "\n";
+                        }
+                        string_mess += "```";
+                        toChannel.send(string_mess);
+                    }
+                    else
+                    {
+                        toChannel.send("*There are currently no songs in the queue. Use $play [song name] to add one to the queue!*");
+                    }
                 }
+            break;
+
+
+            // Plot
+            case 'plot':
+                controller.plot(String_Args);
             break;
             }
         }
